fix some clang issues

+++ src/papilo/presolvers/SimpleSubstitution.hpp
@@ -248,7 +248,7 @@ SimpleSubstitution<REAL>::perform_simple_substitution_step(
          REAL normalized_vals[2] = { num.round( vals[0] ), num.round( vals[1] ) };
 
          // skip if overflowing
-         if( std::max( abs( normalized_vals[0] ), abs( normalized_vals[1] ) ) > std::numeric_limits<int64_t>::max() )
+         if( std::max(  abs( normalized_vals[0] ), abs( normalized_vals[1] ) ) > static_cast<double>(std::numeric_limits<int64_t>::max()) )
             return PresolveStatus::kUnchanged;
          auto res = boost::integer::extended_euclidean(
             static_cast<int64_t>( abs( normalized_vals[stay] ) ),
